backend/requirements.txt
# FastAPI and server
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.10

# OCR
paddlepaddle==3.2.2
paddleocr==2.7.3

# LLM
google-generativeai==0.8.0

# Image processing
pillow==10.3.0
numpy>=1.21.0,<2.0.0
opencv-python-headless>=4.8.0

# Configuration
pydantic-settings==2.3.4
python-dotenv==1.0.1